+++ core/src/org/rexcellentgames/burningknight/entity/creature/mob/desert/Mummy.java
@@ -36,7 +36,7 @@ public class Mummy extends Mob {
 	}
 
 	@Override
-	protected void onHurt(float a, Creature from) {
+	protected void onHurt(int a, Creature from) {
 		super.onHurt(a, from);
 		this.playSfx("damage_mummy");
 	}
