+++ core/src/org/rexcellentgames/burningknight/entity/creature/mob/hall/Clown.java
@@ -91,7 +91,7 @@ public class Clown extends Mob {
 	}
 
 	@Override
-	protected void onHurt(float a, Creature creature) {
+	protected void onHurt(int a, Creature creature) {
 		super.onHurt(a, creature);
 		this.playSfx("damage_clown");
 	}
