+++ core/src/org/rexcellentgames/burningknight/entity/level/entities/chest/Mimic.java
@@ -216,7 +216,7 @@ public class Mimic extends Mob {
 	}
 
 	@Override
-	protected void onHurt(float a, Creature from) {
+	protected void onHurt(int a, Creature from) {
 		super.onHurt(a, from);
 
 		if (!this.found) {
